move routing logic to pwa-kit-react-sdk
